refactor(SignerFactory): encapsulate call to getPublicKey
